* 'feature/versioning-simple-dht' of bitbucket.org:raccc/comsys_challenge_task:
  Use FSPeer and its persistence to access DHT
  Don't sync root directory
  Ignore / while file syncing
  Add path to DHT on complete write
  Remove dht access
  Fix truncated

+++ src/net/f4fs/filesystem/event/listeners/SyncFileEventListener.java
@@ -5,6 +5,7 @@ import java.util.logging.Logger;
 
 import net.f4fs.filesystem.event.events.AEvent;
 import net.f4fs.filesystem.event.events.AfterWriteEvent;
+import net.f4fs.filesystem.util.FSFileUtils;
 import net.tomp2p.peers.Number160;
 
 /**
@@ -33,14 +34,20 @@ public class SyncFileEventListener
 
             // create local non-existing files
             for (String key : keys) {
+                if (FSFileUtils.isRootDirectory(key)) {
+                    // no changes are allowed to root directory
+                    continue;
+                }
+
                 if (afterWriteEvent.getFilesystem().getPath(key) == null) {
-                    System.out.println("Created file again...!?!");
                     // check whether the path is a link, that means key and target are different
                     String foundPath = afterWriteEvent.getFsPeer().getPath(Number160.createHash(key));
                     if (null != foundPath && !key.equals(foundPath)) {
                         // target key is different from source key -> is a symlink
+                        logger.info("Call 'symlink' for target '" + foundPath + "' on path '" + key + "'");
                         afterWriteEvent.getFilesystem().symlink(foundPath, key);
                     } else {
+                        logger.info("Call 'create' for file/dir on path '" + key + "'");
                         afterWriteEvent.getFilesystem().create(key, null, null);
                     }
                 }
@@ -49,6 +56,10 @@ public class SyncFileEventListener
             // remove deleted files / dirs / symlinks / ...
             localPaths.removeAll(keys); // list of all localPaths which are removed in the DHT
             for (String pathToDelete : localPaths) {
+                if (FSFileUtils.isRootDirectory(pathToDelete)) {
+                    // no changes are allowed to root directory
+                    continue;
+                }
                 logger.info("Call removal of element on path '" + pathToDelete + "'. LocalPaths: " + localPaths + ", DHTPaths: " + keys);
                 afterWriteEvent.getFilesystem().unlink(pathToDelete);
             }
@@ -63,5 +74,4 @@ public class SyncFileEventListener
     public String getEventName() {
         return AfterWriteEvent.eventName;
     }
-
 }
